+++ example.cfg.py
@@ -31,7 +31,7 @@ KEY_STORE = GitHubKeyStore()
 # Unlike public keys, the master key ideally ought to be accessible by
 # only Geofront.  Assume you use Amazon Web Services.  So you'll store
 # the master key to the your private S3 bucket named your_team_master_key.
-from geofront.masterkey import CloudMasterKeyStore
+from geofront.backends.cloud import CloudMasterKeyStore
 from libcloud.storage.types import Provider
 from libcloud.storage.providers import get_driver
 
@@ -45,13 +45,24 @@ MASTER_KEY_STORE = CloudMasterKeyStore(driver, container, 'id_rsa')
 # but you'll get the list dynamically from EC2 API.  Assume our all
 # AMIs are Amazon Linux, so the usernames are always ec2-user.
 # If you're using Ubuntu AMIs it should be ubuntu instead.
-from geofront.remote import CloudRemoteSet
+from geofront.backends.cloud import CloudRemoteSet
 from libcloud.compute.types import Provider
 from libcloud.compute.providers import get_driver
 
 driver_cls = get_driver(Provider.EC2_US_WEST)
 driver = driver_cls('aws access id', 'aws secret key')
 REMOTE_SET = CloudRemoteSet(driver, user='ec2-user')
+
+# Suppose your team is divided by several subgroups, and these subgroups are
+# represented in teams of the GitHub organization.  So you can control
+# who can access each remote by specifying allowed groups to its metadata.
+# CloudRemoteSet which is used for above REMOTE_SET exposes each EC2 instance's
+# metadata as it has.  We suppose every EC2 instance has Allowed-Groups
+# metadata key and its value is space-separated list of group slugs.
+# The following settings will allow only members who belong to corresponding
+# groups to access.
+from geofront.remote import GroupMetadataPermissionPolicity
+PERMISSION_POLICY = GroupMetadataPermissionPolicity('Allowed-Groups')
 
 # Geofront provisions access tokens (or you can think them as sessions)
 # for Geofront clients.  Assume you already have a Redis server running